backend/requirements.txt
django>=6.0,<6.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
channels>=4.1
daphne>=4.1
django-cors-headers>=4.3
